+++ R/gg_hbar.R
@@ -324,7 +324,6 @@ gg_hbar <- function(data,
 #' @param y_zero_line For a numeric y variable, TRUE or FALSE of whether to add a zero reference line to the y scale. Defaults to TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE.   
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles.
 #' @param col_labels A vector of colour labels.   
-#' @param col_labels_dp For numeric colour variables and where col_labels equals NULL, the number of decimal places. Defaults to 1 for "quantile" col_method, and the lowest dp within the col_cuts vector for "bin".
 #' @param col_legend_ncol The number of columns in the legend. 
 #' @param col_legend_nrow The number of rows in the legend.
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
@@ -400,7 +399,6 @@ gg_hbar_col <- function(data,
                         y_zero_line = NULL,
                         col_cuts = NULL,
                         col_labels = NULL,
-                        col_labels_dp = NULL,
                         col_legend_ncol = NULL,
                         col_legend_nrow = NULL,
                         col_method = NULL,
@@ -506,7 +504,6 @@ gg_hbar_col <- function(data,
       }  
       col_cuts <- stats::quantile(col_var_vctr, probs = col_cuts, na.rm = TRUE)
       if (anyDuplicated(col_cuts) > 0) stop("col_cuts do not provide unique breaks")
-      if(is.null(col_labels_dp)) col_labels_dp <- 1
     }
     else if (col_method == "bin") {
       if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr, col_pretty_n)
@@ -514,13 +511,12 @@ gg_hbar_col <- function(data,
         if (!(dplyr::first(col_cuts) %in% c(0, -Inf))) warning("The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)")
         if (dplyr::last(col_cuts) != Inf) warning("The last element of the col_cuts vector should generally be Inf")
       })
-      if(is.null(col_labels_dp)) col_labels_dp <- sv_max_dp(col_cuts)
     }
-    
+
     data <- data %>% 
       dplyr::mutate(dplyr::across(!!col_var, ~cut(.x, col_cuts, right = FALSE, include.lowest = TRUE)))
     
-    if(is.null(col_labels)) col_labels <- sv_numeric_bin_labels(col_cuts, col_labels_dp)
+    if(is.null(col_labels)) col_labels <- sv_numeric_bin_labels(col_cuts, sv_max_dp(col_cuts))
     
     col_n <- length(col_cuts) - 1
     if (is.null(pal)) pal <- pal_viridis_reorder(col_n)
@@ -1059,7 +1055,6 @@ gg_hbar_facet <- function(data,
 #' @param y_zero_line For a numeric y variable, TRUE or FALSE of whether to add a zero reference line to the y scale. Defaults to TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE.   
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles.
 #' @param col_labels A vector of colour labels.   
-#' @param col_labels_dp For numeric colour variables and where col_labels equals NULL, the number of decimal places. Defaults to 1 for "quantile" col_method, and the lowest dp within the col_cuts vector for "bin".
 #' @param col_legend_ncol The number of columns in the legend. 
 #' @param col_legend_nrow The number of rows in the legend.
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
@@ -1135,7 +1130,6 @@ gg_hbar_col_facet <- function(data,
                               y_zero_line = NULL,
                               col_cuts = NULL,
                               col_labels = NULL,
-                              col_labels_dp = NULL,
                               col_legend_ncol = NULL,
                               col_legend_nrow = NULL,
                               col_method = NULL,
@@ -1254,7 +1248,6 @@ gg_hbar_col_facet <- function(data,
       }  
       col_cuts <- stats::quantile(col_var_vctr, probs = col_cuts, na.rm = TRUE)
       if (anyDuplicated(col_cuts) > 0) stop("col_cuts do not provide unique breaks")
-      if(is.null(col_labels_dp)) col_labels_dp <- 1
     }
     else if (col_method == "bin") {
       if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr, col_pretty_n)
@@ -1262,13 +1255,12 @@ gg_hbar_col_facet <- function(data,
         if (!(dplyr::first(col_cuts) %in% c(0, -Inf))) warning("The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)")
         if (dplyr::last(col_cuts) != Inf) warning("The last element of the col_cuts vector should generally be Inf")
       })
-      if(is.null(col_labels_dp)) col_labels_dp <- sv_max_dp(col_cuts)
     }
-    
+
     data <- data %>% 
       dplyr::mutate(dplyr::across(!!col_var, ~cut(.x, col_cuts, right = FALSE, include.lowest = TRUE)))
     
-    if(is.null(col_labels)) col_labels <- sv_numeric_bin_labels(col_cuts, col_labels_dp)
+    if(is.null(col_labels)) col_labels <- sv_numeric_bin_labels(col_cuts, sv_max_dp(col_cuts))
     
     col_n <- length(col_cuts) - 1
     if (is.null(pal)) pal <- pal_viridis_reorder(col_n)